backend for apply loan merged

+++ springapp/src/main/java/com/examly/springapp/controller/AdminController.java
@@ -31,27 +31,27 @@ public class AdminController {
 		return calladminser.getAll();
 	}
 	@DeleteMapping("/displaycustomers/{id}")
-	public String deleteCustomer(@PathVariable("id") int id)
+	public String deleteCustomer(@PathVariable("id") Long id)
 	{
 		return calladminser.deleteCustomer(id);
 	}
 	@GetMapping("/displaycustomers/{id}")
-	public LoanModel getCustomer(@PathVariable("id") int id)
+	public LoanModel getCustomer(@PathVariable("id") Long id)
 	{
 		return calladminser.getCustomer(id);
 	}
 	@PutMapping("editCus/{id}")
-	public LoanModel changeStatus(@PathVariable("id") int id,@RequestBody LoanModel t)
+	public LoanModel changeStatus(@PathVariable("id") Long id,@RequestBody LoanModel t)
 	{
 		return calladminser.editCustomer(t,id);
 	}
 	@GetMapping("approve/{id}")
-	public boolean approve(@PathVariable("id") int id)
+	public boolean approve(@PathVariable("id") Long id)
 	{
 		return calladminser.approve(id);
 	}
 	@GetMapping("reject/{id}")
-	public boolean reject(@PathVariable("id") int id)
+	public boolean reject(@PathVariable("id") Long id)
 	{
 		return calladminser.reject(id);
 	}
